requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.3.0
attrs==19.3.0
cachetools==4.1.1
certifi==2019.11.28
chardet==3.0.4
chromedriver-autoinstaller==0.2.0
colorama==0.4.1
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.8
importlib-metadata==0.23
more-itertools==7.2.0
MouseInfo==0.1.3
oauthlib==3.1.0
packaging==19.2
Pillow==7.2.0
pluggy==0.13.0
protobuf==3.13.0
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==2.4.2
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
pytest==5.2.2
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.6
wcwidth==0.1.7
zipp==0.6.0
